524 bug - %2F slashes translation behind mount point (unit test)
New test getDataWithSlashesBehindMountPoint() tests whether URI identifier
which contains slashes encoded as %2F (e.g. in key part of URI) is translated to identifier with simple slash
characters (/).

It is checked whether InstanceIdentifier instance is created with simple
slashes. The instance is then used as input parameter of method
BrokerFacade.readOperationalDataBehindMountPoint() to get data from
datastore.

If %2F are correctly replaced with / test is succesful.

Change-Id: I131bda42fb568af23d6a1b0bbb72c505db8748e3
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>