Skip from the surefire execution plugins that can step over guava feet
Seems like the com.google.common.collect.* package has been reused in
several jars like:

- guava
- google-collection
- pax-url-aether

Now the netconf-it in the classpath of contains via transitive the 3
of them but only guava is usable by yang-bindings used by the netconf
client test. The presence of the 3 cause unreliability on the test
pass based on who comes first in the classpath of surefire. The fix is
to make sure surefire ignore the other two libraries except guava.

Change-Id: I77260989a19f60db94286434444530bacb1947d4
Signed-off-by: Giovanni Meo <gmeo@cisco.com>