Trigger a GC once initial configuration has been pushed
This introduces a single run of GC once we complete the initial
configuration.

The reasoning behind that is that we generate a fair amount of metadata
which we are likely to retain unmodified for extended periods of time.
We take care to store it in both memory and CPU efficient format, which
unfortunately requires us to do some copying around to compact it,
leaving a sizeable chunk of garbage on the heap.

Change-Id: I81af7d317e3b6f82a431e78b5fb439d8cb5f1997
Signed-off-by: Robert Varga <rovarga@cisco.com>