Bug 3999: Create internal service to access restconf
There are use cases for invoking restconf from internal code. However
issuing an HTTP request is problematic as one would need to know the
credentials and scheme (http or https).

So I added a JSONRestconfService interface and implementation with CRUD
methods that call the JSON readers/writers and RestconfImpl internally.
The implementation is advertised as an OSGi service via the config
system for consumption by clients.

I only added a service for JSON - an XML service could be added as well
later.

Change-Id: Ia3753846756ace29bc28aa0a1fe9d474a60f80ce
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>