Fix failing LISP tests due to implementation changes
When returning a mapping on the RESTCONF interface, the mapping service
crafted the same object as the one sent on the southbound, according to
the LISP protocol (which is a Map-Reply with no locator records, action
bit: forward native). However, getting a reply on the northbound with
HTTP code 200 when actually there is no mapping in the database is
confusing, and not what a NB user would expect. The behavior of the
mapping service was changed in Beryllium to return HTTP 404 when no
mapping is present.

This commit accounts for the above change, and additionally factors out
common variables and keywords into Variables.py and a new LISP resource
respectively.

Change-Id: Ia7c8fd2adb2408f2b1c17865bf5512307dc1e635
Signed-off-by: Lorand Jakab <lojakab@cisco.com>