Moved the karaf logging keywords to KarafKeywords
This change defines API-like keywords to be used in suites.
Actual implementation may change in subsequent changes.

Original idea an main implementation by: Jozef Behran
Finishing touches by: Vratko Polak.

The keywords are moved with as little modifications as possible
(added code to preserve the active SSH connection across
these calls, renamed "Connect To Controller Karaf" to "Open
Controller Karaf Console On Background" and added requirement
that "Open Controller Karaf Console On Background" is used
prior to any "Log Message To Controller Karaf"),
and without any changes to the rest of the KarafKeywords
because these are used in quite a number of other suites.
Merging this new code with the existing KarafKeywords codebase
is a long and involved work that is better left to another commit.

Additionally, this change updates the BGP Functional test
to use the new keywords. This serves as an example about how
to use these keywords and additionaly provides an additional
verification that the keywords are not broken because the
integration job will rerun the test which will fail,
exposing the breakage. This change prompted for creating yet
another resource (SetupUtils.robot) with keywords that are
used in the [Setup ...] and [Teardown ...] sections of the
suite/test. These are going to be used in further test suites
so it is better to create them now to avoid code duplication
(actually code triplication as there are 2 more tests being
created that need them).

Change-Id: I5df655b9532aeeee377d89e21743e74f1050498e
Signed-off-by: Jozef Behran <jbehran@cisco.com>
Signed-off-by: Vratko Polak <vrpolak@cisco.com>