Update patch set 22
Patch Set 22:

> (1 comment)

So we are still using Jersey 1.X (I think you originally pushed 2.X patch), which uses web.xml to define the Jersey Servlet.  Before Shiro, we used a ContainerRequestFilter, which is a Jersey 1.X specific filter specified in the init-param of the Application.  Shiro is not Jersey dependent (in fact, it can work with anything we want even non-jersey in the future if that is what we decide), and Shiro uses the more generic javax.servlet.Filter abstraction.  Well in web.xml, you are allowed to specify a ContainerRequestFilter but never have it in the classpath (Jersey will just ignore it).  This is how the noauth feature originally worked (it just didn't import the bundle that contains the AAA TokenAuthFilter) so the Filtering mechanism was never called (in fact, if you install odl-restconf, stop karaf, then start again and install odl-restconf-noauth instead with the old approach, you will still have AAA enabled since the bundle is cached in the data/ directory).  However, with Jersey and javax.servlet.Filter (new proposed approach), one MUST have all specified Filter(s) in the classpath, or the Jersey filter will not even start (fails out with an exception and you can't hit restconf period).  Thus, the proposed solution is to make AAAFilter disabled by default, and only enable if odl-shiro-act bundle is installed (which is done only in odl-restconf), but to include the dependencies regardless.  The mechanism is similar, but needed in order to achieve the existing behavior.

Do you have an alternate idea?

Patch-set: 22
Label: Verified=0