Update patch set 1
Patch Set 1:

If you look at the line right before the switch statement:
String alignmentString = EXIParameters.EXI_PARAMETER_BIT_PACKED;
This sets the default value right before the switch.

You can leave alignmentString uninitialized and initialize it to the
default value with something like:
default with something like this:
            LOG.warn("Unexpecte value in EXI alignment type: {} , using default value", exiOptions.getAlignmentType());
        case bitPacked: {
            alignmentString = EXIParameters.EXI_PARAMETER_BIT_PACKED;
            break;
        }

Note that with this you don't want to break in the default case since you want it to fall through to bitPacked.

Patch-set: 1