Bug 4577 Allow specification of a distinct schema cache directory per netconf device
A leaf is added to odl-sal-netconf-connector-cfg.yang to allow specification of
a schema cache directory relative to the "cache" directory.  The leaf defaults to
"schema", so the default directory for loaded yang files is "cache/schema".  It
is useful to specify a distinct cache directory per netconf mount to avoid potential
model conflicts caused by revisionless import statements.

Change-Id: Id930932aa4ff17a7338ddf96e87557b3949b4204
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>