Update patch set 5
Patch Set 5:

> we may be deprecating mount through CSS in future

In my eyes, the main added value of the CSS way is automatic creation of controller-config device that allows user to reconfigure ODL in runtime via Restconf. I do not think there will be an alternative for this in Boron, so I expect deprecation in Carbon and removal in Nitrogen.

> I did not see a need to abstract common code and pull into a separate class

Perhaps good enough for Beryllium, and I believe devs can do the de-duplication in Boron (and Be-SR1). (But I may be wrong as I am not a dev around here.)

Patch-set: 5