Bug 4948: Catch RuntimeException from ensureParentByMerge
Made the calls to PUT and POST more robust. Caught the RuntimeException from
 the merge call in the EnsureParentsByMerge. If this fails then started
a new transaction and continue with the original PUT. This ensures when a device
is unable to handle the empty list/container merge it will still succeed with the
original request. Similar change has been made for POST.
THis bug was fixed in stable Lithium. I am pushing the same fix on Beryllium. This fix makes the
PUT and POST calls more robust on Netconf devices by ignoring the errors resulting from mergeParents
if a device is not able to handle it.

https://git.opendaylight.org/gerrit/#/c/32457/ was the gerrit ticket for the Li checkin

Change-Id: Id18a97857e46f44014dbd61cbfa86285457a2905
Signed-off-by: Balaji Varadaraju <bvaradar@brocade.com>