Bug 4948: Catch RuntimeException from ensureParentByMerge
Made the calls to PUT and POST more robust. Caught the RuntimeException from
 the merge call in the EnsureParentsByMerge. If this fails then started
a new transaction and continue with the original PUT. This ensures when a device
is unable to handle the empty list/container merge it will still succeed with the
original request. Similar change has been made for POST.
This bug was fixed in stable Lithium. I am pushing the same fix on Beryllium.

Added FIXME statements to figure out and catch specific Runtime exceptions
thrown by Netconf instead of RuntimeException.

Change-Id: Id18a97857e46f44014dbd61cbfa86285457a2905
Signed-off-by: Balaji Varadaraju <bvaradar@brocade.com>
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>