Bug 4577 Allow specification of a distinct schema cache directory per netconf device
First iteration done here https://git.opendaylight.org/gerrit/#/c/30163
introduce a BUG: sideloaded schemas, manually added in cache/schema, and
specified in the netconf connector, where not being taking in account
in the specified cache folder.

This patch make sure sideloaded schemas are copied in the specified
directory, and make sure those schemas are correctly processed by
FilesystemSchemaSourceCache.

The reason why we create the directory and add the schemas in there
before create the instance of FilesystemSchemaSourceCache is because
FilesystemSchemaSourceCache loads the schemas from the dir only once,
during the instantiation.

Change-Id: I6a5a6b23fc56bcc507734a99e48de88597b7ae2a
Signed-off-by: adetalhouet <adetalhouet@inocybe.com>