Update patch set 3
Patch Set 3: Code-Review-1

This should really be handled in another way. You should be able to register the sources in the default cache folder as a fallback into the custom cache repository albeit with a lower priority. This won't conflict if there already is a schema present in the custom repo directory.

Patch-set: 3
Label: Code-Review=-1