Fix netconf port variables
There are two netconf ports in ODL, one is MDSAL and the
other one is the config subsystem. However there was only
one Netconf port variable, confusingly named ODL_NETCONF_PORT
which pointed to the MDSAL port. Fix that by putting the
name of the subsystem behind the port into the name of the
variable and also create the other variable which for some
reason was still missing.

Change-Id: I526f1a314a7a3b1414f333f9b161f146e5e88d87
Signed-off-by: Jozef Behran <jbehran@cisco.com>