Find the most recent notification
If multiple revisions for same notifications are present
it will pick the most recent one. It will iterate through
all of the notifications and it will pick the one with most
recent revision date. StripNotification function is also moved
to NetconfMessageTransformUtil class together with
EVENT_TIME_FORMAT which is needed for this function.

Change-Id: Ibd9fe800d75f877ead03162b42895b322288b62e
Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>