Fix bug 4548:
A leaf node named sequentially-send-request is increased in grouping
netconf-node-connection-parameters in netconf-node-topology.yang file.
Type of this leaf node is boolean, and its default value is true.

When its value is true, the netconf device communicator will send requests
to the device sequentially. That is, only after the communicator receives
the response of the request, it will send the next request to the device.
When its value is false, the communicator will send requests to the device
concurrently.

The default value of this leaf node is true, and when creating a new
netconf device in ODL by REST API, the value of this leaf node
corresponding to this new netconf device can be set to false if the device
supports processing netconf requests concurrently.

Change-Id: Ie9d755de23a650a0465b540e538c6f537db87a7f
Signed-off-by: Nanfei Chen <chen.nanfei@h3c.com>