Update patch set 5
Patch Set 5:

> How can you be sure that all the data that was added by any other
 > application to the node is still relevant once the node is
 > reconnecting?
The fundamental thing here appears to be in understanding what are the add/update semantics in terms of the netconf node events. 
It appears reasonable to consider that a reconnecting node was actually disconnected first, and thus its state removed - a reconnect should then be via an add then, while a disconnect via delete, at least in operational state.
It does appear however that you're thinking more in terms of a reconnecting netconf node as being an update which is likely the source of confusion. Why would this semantic be more useful than the former?

 > Also the netconf server capabilites or anything else might change
 > on reconnect and we would end up with inaccurate data.
 > 
 > The easiest solution would be to just delete the capabilities
 > before updating the state but that still would not solve dangling
 > data that some other applications might write here.

Patch-set: 5
Reviewer: Gerrit User 2106 <2106@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0