Bug 5708: Schemaless netconf mount point
Schemaless mount point can be used to connect to devices
which doesn't provide their schemas. In that case,
both rpc input and output must be of anyxml type.

Data broker and transactions works with anyxml too. Read transaction path
Yang instance identifier is transformed to xml filter structure
used in get-config rpc. Resulting data are selected from rpc-reply
according to path, wrapped in data element(to be able to return lists)
and returned as anyxml. Write transaction path is transformed to
xml structure. Data are appended to the last element of this structure.

This is done by SchemalessRpcStructureTransformer. It is strategy
used in NetconfBaseOps class. For standard mount points
NetconfRpcStructureTransformer is used.

Change-Id: I38c70fc6b676941ef6d731a52a6b12ae7c10eec6
Signed-off-by: Andrej Mak <andmak@cisco.com>