Bug-6346: Allow over-ride of non-module capabilities
This feature will allow user to override non modle based capability like
:writable-running or :candidate capabilities of the device. It would be useful
to be able to specify the none model capability to use in the node configuration
using something similar to yang-additional-capabilities to add or over ride
the capabilities announced by the device. This is a situation where manual netconf
can work around the bad hello capabilities
exchange but the ODL netconf adaptor can not.

Change-Id: I1ba4ac1ca84a528434658e8af7c3db981146edb6
Signed-off-by: Om Prakash <op317q@att.com>