Update patch set 4
Patch Set 4:

> Patch set 2 does not trigger netconf-distribution-check-beryllium,
 > but patch set 3 triggered. Why?

It is because patchset 2 did not have any code changes (only the commit message was changed). The distribution-check job is configured to ignore a patchset if no code was changed since the last patchset. If you think this is incorrect and should be tested on every run please contact integration-dev.

Patch-set: 4