Update patch set 4
Patch Set 4:

> I don't know, if someone is actually using it.. But what if someone does? 

If someone does that, it means he knows how to get around things within ODL, e.g. he would be able to manage/adapt to this change :)


> Shouldn't we say somewhere, that this is no longer supported?

Fair enough, I will deprecate it and remove this is Nitrogen then. So this remains backward compatible for Carbon with a deprecation notice.

Patch-set: 4