Add blueprint wiring for netconf-client
- move netconf-config-dispatcher to netconf-client because the *Module
is creating a proxy that needs to share the same classpath as the netconf-client
as it is the bundle providing its implementation and some of its classes
aren't public, but either package-private or protected.

- remove netconf-config-dispatcher from everywhere, but keep the models
there for now as the USC project is dependent on them. To avoid having a flag
day, we keep the models here until https://git.opendaylight.org/gerrit/#/c/46621/
is merged. That patch will be eligible to be merged only once this current
patch is merged.

Change-Id: I2b7007115bf7228ceac6db5946d4e0b859d8ecc6
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>