Update patch set 2
Patch Set 2:

> when stream is not default, the value is using the input value

Can you point me to the line of code in the class please? I'm having trouble to understand, as in this patch https://git.opendaylight.org/gerrit/#/c/47460/ I can remove the stream param because it is clearly unused. I must be missing something so if you can please point it out that would be appreciated.
(I'm not saying my patch is good tough)

Patch-set: 2