Update patch set 16
Patch Set 16:

> How you could fix that deadlock by removing CSS and yang files? Two blueprint threads deadlocked, I don't think that's anyhow related to CSS.

Are you saying you're still facing the issue when you grad the whole patch chain>

Here is my explanation
Currently, we are still instantiating module and creating a proxy for the registered service in the *Module#createInstance(). And that proxy service is used by config system for other modules. But the origin of the service is coming from the blueprint, not CSS.
So on shutdown, config modules are the first to get destroyed, so the proxies get destroyed, hence the blueprint container depending on those proxy are waiting to retrieve the proxy to correctly #close. This is true in NetconfOperationServiceFactoryListener#close()

Patch-set: 16