Update patch set 2
Patch Set 2:

> Not sure why validate-autorelease passed but it shouldn't have. The version in the pom.xml file for mdsal is incorrect.

Well, at the time the patch was cherry-pick and validated, the version was accurate. But in-between the version bump patch came in.

Timeline is:
c/47394/ Oct 24 <-- that patch verified and OK
c/47918/ Nov 3  <-- version bump
c/47394/ Nov 10 <-- that patch merged

The problem is that patch was not rebased. And I don't have any clear answer whether the release team should rebased patches including pom modification having internal version of ODL in the diff, or if it is the committer that should be more cautious at that time in the release.

Patch-set: 2