Update patch set 3
Patch Set 3:

We thought about user credentials. We want to address this issue ( two way encryption, key based auth, certificate based auth etc) when there is a comprehensive strategy in ODL for the same. Today for any user credentials (NETCONF is just one example) there is no comprehensive way to obfuscate the sensitive information. SO we will address this at a later point of time.

Patch-set: 3