Update patch set 3
Patch Set 3:

I would still argue for a better insight on this. For example for two way encryption which algorihms clients must use, what's the strategy for selecting a SALT? I am thinking that should help all of the applications and not each one implementing their own strategies and code. Is there a central place to do this? is AAA a good candidate for this?

Patch-set: 3