Clean ControllerContext up
ControllerContext is supposed to be an immutable context, which has always
a global instance, which is updated as needed.

Current code mutates the context in an uncoordinated fashion, which means
it can easily end up being inconsistent if a SchemaContext change occurs
while a request is being executed.

Add FIXME markers to issues to be solved in future and clean up various
aspects of the file:
- make sure the URI encoding charset is available
- use lambdas wherever possible
- fix keyword order to match JLS
- trim long lines

Change-Id: I210d5ee6bff7ea9d4853daa7f36411da77388c69
Signed-off-by: Robert Varga <rovarga@cisco.com>