Update patch set 7
Patch Set 7:

Lori: thanks.

javax.annotation is JSR305, just as we discussed on the MD-SAL call. jsr305 used to be pulled in by aaa, hence netconf would get it.

Have we decided who is supposed to provide it? If odlparent, we should add it as a boot feature in karaf-empty...

Patch-set: 7