Update patch set 12
Patch Set 12:

We will actually have to work more on aaa features to get this completely sorted out -- that osgi.enterprise dependency is a huge problem, as it ends up restarting the world. The correct approach there is to depend on karaf assemblies etc -- but we will take care of that once this is merged.

Patch-set: 12