Update patch set 4
Patch Set 4:

> Bumping this to see if we can get some further feedback from
 > project committers/contributors.  Thanks!

I would also say that if we wanted to approach this as a Beta contribution, I'd be fine with allowing the current auth mechanism with the assumption that we would immediately open a bug to track securing the credentials.  For reference, AAA does expose a standard encryption service (aaa-encryption-service), which provides encryption and decryption methods based on a common set of configured parameters.  This is something I would recommend using for encryption in the future.

Patch-set: 4