Bug 7869 - handle shcema without namespace.
currently when the ODL connects to a device and recieves a schema
without a namespace it disconnects.
This fix allows the ODL to ignore schemas without namespaces, but still
maintains the connection and handles the rest of the schemas which do
have a namespace.
To test this i removed the namespace from one of the schmas, and
decreased the number of expected schemas by one, so we can see that we
manage to import all of the schemas, and just ignore the problematic
one.

Change-Id: I9134fce17a85a2b5111e5acd0c63516f4c356db8
Signed-off-by: David Goldberg <gdavid@hpe.com>