Reuse filter XPaths
Speed up notification filters by making sure we pre-compile the supplied
filter into an XPathExpression and reuse it for all notifications. Since
XPathExpression is not thread-safe, we synchronize on it to ensure we
are not accessing it concurrently.

Also reuse the DocumentBuilderFactory during parse phase of the XML. This
is not as efficient as it should be, though.

Change-Id: I7ee8c4d476d971ec505769908e97a82070a9ea88
Signed-off-by: Robert Varga <rovarga@cisco.com>