Eliminate ListenerConstants
These constatns are used only in AbstractNotificationData, inline them
in that class.

lso convert to using Java 8 Time API to fix thread-safety issues with
SimpleDateFormat. Since we need to choose an appropriate API, we choose
to report instants in UTC, so we do not have a dependency on the concept
of the local time zone.

Change-Id: I68153d525960d47e6ff60634df976d472beb2176
Signed-off-by: Robert Varga <rovarga@cisco.com>