Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

I didn't really do a thorough review just yet, but please turn on checkstyle on the impl modules, we want to eventually turn it on the whole netconf project but didnt have the manpower to fix all the violations, but it's a good idea to have it on for new contributions. You can create a new patch that before this one which turns it on and fixes all the violations there.

Patch-set: 3
Label: Code-Review=-1