Update patch set 3
Patch Set 3:

Uh, not sure what to say. This code base is done. It is being ported upstream, meaning this is not the inception of a project. Turning on things like check style can result is hundreds or thousands of issues; try using it against your favorite upstream code and you will be (rightly) horrified. We are going to continue some work on this post-Carbon. Can I open a defect to enable check style and fix those issues as a separate bug? Is that acceptable? FWIW we are trying to beat out the M5 code freeze and this request could kill that.

BTW thanks for all the time reviewing the code; it is better code for your efforts.

Patch-set: 3