Update patch set 5
Patch Set 5:

> I didn't really do a thorough review just yet, but please turn on
 > checkstyle on the impl modules, we want to eventually turn it on
 > the whole netconf project but didnt have the manpower to fix all
 > the violations, but it's a good idea to have it on for new
 > contributions. You can create a new patch that before this one
 > which turns it on and fixes all the violations there.

Do we have this documented somewhere? Even if they are (which I don't think they are), that adds an annoying extra thing to find the requirements or be informed of them. 

If we're going to reject patches for checkstyle violates, we _really_ need to have it turned on in the code where people can easily find it in an automated way.

Otherwise, it's actively hostile to new developers to have them navigate a minefield of unknown requirements.

Patch-set: 5
Label: Verified=0