Bug 8074 - actor name [test-device] is not unique!
When creating a new actor with same name as an old one, it might fail
becuase old actor might not be completely gone at this point.

This creates a slaveActor which will never be closed, but instead
every time that mountpoint is closed or master is changed, it would
update NetconfNodeActor data and keep the same actor.

Only time that actor is closed is when node was deleted. Or slave
became a master

The name of the actor is changed to device id plus timestamp,
which should definitly create a actor with new name.

Change-Id: I6d10db7b8902048632e336fb5a9701a797b363c6
Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>