Bug 8032 - Initialization in sal failed, disconnecting from device
It throws a MissingMountpointServiceException if Master was closed before
he was able to create a mountpoint

It seemes to be a problem only because it is logged as an error.
When I compared times in suit and karaf logs in which error appears,
it doesn`t cause any robot test to fail in that suit.

This error should appear when new master is assigned while old one is
still creating a mountpoint and therefore DOMMountPointService is set
to null

Change-Id: I9aa9064752043819bc73043f7b667b4ea2c0fd0a
Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>