Update patch set 9
Patch Set 9:

I did already look for such a tool. I found the google check style sheet and imported it into IntelliJ. Unfortunately it only handles a very tiny fraction of the issues that check style raises, so it is not viable as a tool for converting all of the Netconf project, which is what I was looking for. There is also a tool which will show check style violations "in real time" in IntelliJ, but again nothing that auto-corrects.

Patch-set: 9