Update patch set 9
Patch Set 9:

So are we good to merge? Is the presence of a space in front of a parenthesis really what's holding this up? What if I submit a defect against "call home" for clearing up all check-style issues and get that done in the next code push (estimated 2-3 months).

Patch-set: 9