Update patch set 10
Patch Set 10:

Hi, Alex. Those docs are still accurate but now they are incomplete (we added more). We actually have some integration tests we'd like to commit "real soon now". I'm also assuming that we need to say that this is an experimental feature. I will correspond via email with you on how to update and post new docs.

Patch-set: 10