Update patch set 2
Patch Set 2:

Wow, 4am in the morning - impressive! Actually thanks for the review. I assume that you are the one to click the merge button, as I am not a committer. I assume this will go onto master and that I will have to create a new patch to cherry pick this onto the call home on stable/carbon, right? (I didn't see this workflow described on the wiki)

Patch-set: 2