BUG-5581: batch read check
When posting data we are executing potentially a lot of reads,
which can hurt if the backend is not residing on the node which
is executing the operations.

Fix this by issuing read requests in a batch and then collecting
the results, which means we will amortize request latency.

Change-Id: Ibf5628f6fdf56b7845b4a4623255ff4d417c4546
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 85b7c8994203dd04c380f1917f2c5d2194d89e36)