Separate slave and master actor in singleton
It is clearer that what is going on in master and what in slave.
It also saves some memory because we don`t have to initialize some
objects in slave and some in master.

ReadTxActor was created every time in NetconfNodeActor no matter if
it is slave or master but we don`t need that in slave actor

Change-Id: I955e4fb0f35b1f55faa193ad13557451056201a3
Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>