Bug 8311 - Apidoc: Incomprehensible 500 id model is wrong Bug 8266 - Apidoc explorer is broken after installing Boron SR3
Depending on path if it is relative or absolute module has to be
resolved different ways. If it is relative path it will have to
be resolved through DataSchemaNote which returns namespace and
revision of module in which node is situated. If it is absolute
path we need TypeDefinition which returns a namespace and
revision of original module.

Change-Id: I522ea66383ba1d65e6674d5e8ff67ec31e019303
Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>