Implementation of URL Capability
The NETCONF peer has the ability to accept the <url> element in the source and <target> parameters. The capability is further identified by URL arguments indicating the URL schemes supported. The capability is implemented according to RFC 6241, section 8.8 The capability provides only the File as part of the URL. As is advertised only urn:ietf:params:netconf:capability:url:1.0?scheme=file . File is absolute path to file in the file-system. Access to files are limited. The limitation is specified by folders where ODL has read only access and the folders with read-write access. The update is comprised of:

1. Modifies the <copy-config>. The operation accepts the <url> element as the value of the source and the <target> parameters. The file that the url refers to contains the complete datastore, encoded in XML under the element <config> in the "urn:ietf:params:xml:ns:netconf:base:1.0" namespace.
2. Modifies the <delete -config>. The operation accepts the <url> element as the value of the <target> parameters.
3. Modifies the <validate>. The operation accepts the <url> element element as the value of the source parameter.

Change-Id: I8ee3b2e2455718369efe72acf672d993adfc5fca
Signed-off-by: lbeles <lbeles@frinx.io>