Update patch set 10
Patch Set 10:

> since md-sal is transactional, the only error-option that can be
 > implemented is rollback-on-error. Why are there different options
 > allowed and how do they affect the behavior?

Just error-option is not possible. Because the datastore candidate is one MD-SAL transaction in Netconf Md-SAL. And If edit-config had a rollback-on-error behavior then candidate datastore had to contain a list of MD-SAL transactions. Then when you call a commit, you have to commit the whole list of transaction which are not atomic together. Sollution is rewritten the candidate to real datastore which would be independent from running datastore. But this is a big task.

Patch-set: 10
Label: Verified=0