Update patch set 1
Patch Set 1:

> These were not deprecated as a part of the checkstyle patch if you
 > actually look closely. Deprecation happened in https://git.opendaylight.org/gerrit/#/c/48369/
 > which i guess was a part of moving to the new restconf
 > implementation.

I missed that. That still doesn't explain why it was deprecated and there's no comment indicating the reason. What is the intent? Deprecation usually implies removing it however, as I mentioned, it's a useful interface and is used (of course that's why it was added in the first place). You can't just simply remove it. If the issue is that it uses the old implementation then it should be converted to the new  implementation and not deprecated.

Patch-set: 1
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0