Update patch set 1
Patch Set 1: Code-Review+2

Alright this was some confusion in regards what this was supposed to be on the restconf guys's side. Keep in mind when/if old restconf implementation is deprecated this will be deprecated along with and a replacement should be created that is routed to the new implementation.

Patch-set: 1
Label: Code-Review=+2