Remove UriInfo from JSONRestconfService API methods
https://git.opendaylight.org/gerrit/#/c/48369/ added a UriInfo parameter
to several JSONRestconfService API methods in Carbon. However this was an
undocumented API change which will break downstream users. Further,
there is no available implementation of UriInfo so each user would have
to provide an implementation which would pretty much render the API
unusable. We do not need to expose UriInfo in this API so I've remove the
parameters.

We'll need to cherry-pick to stable/carbon for SR1 to minimize breakage.

Change-Id: I74285413c2a3a92510398a57ea82567b646c6d27
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>