Update patch set 26
Patch Set 26:

If I'm looking at this right this still wont be 100% secure, you still will be able to hit a window when you could grab the unencrypted password.
Whats happening is you let the user write the password into the ds(you can hit a window when its visible unencrypted in the datastore) and then you grab the pw, encrypt it and rewrite it in the datastore.
To avoid this we would have to avoid letting the user write the password directly into the ds, maybe a custom rpc?

Patch-set: 26