Update patch set 26
Patch Set 26:

> If I'm looking at this right this still wont be 100% secure, you
 > still will be able to hit a window when you could grab the
 > unencrypted password.
 > Whats happening is you let the user write the password into the
 > ds(you can hit a window when its visible unencrypted in the
 > datastore) and then you grab the pw, encrypt it and rewrite it in
 > the datastore.
 > To avoid this we would have to avoid letting the user write the
 > password directly into the ds, maybe a custom rpc?

Can we do that as a follow-up and not let the perfect be the enemy of the good? I agree we should figure it out, but that would also not be backward compatible, which probably matters some.

Patch-set: 26
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0