Update patch set 1
Patch Set 1:

Hi Tomas, Thanks for the review.

I will upload a second patch that adds a getter for the deviceRpc in KeepaliveDOMRpcService.

But I think we should keep KeepaliveDomRpcService a static nested class because it uses other nested classes in KeepaliveSalFacade: ResetKeepalive and RequestTimeoutTask, both of which are private.

Patch-set: 1
Label: Verified=0