BUG-8697: select correct transformer for schemaless netconf mounts
Getting or editing the configuration of a schemaless-mounted netconf device
from the code of an ODL application results in a NullPointerException.

This patch fixes the NetconfBaseOps constructor so that the RPC structure
transformer dedicated to schemaless mount points is actually selected.

Change-Id: Ib0b2d1c9910613f4354bdcb4f1c56e9be36a95b2
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@b-com.com>
(cherry picked from commit 32621d57701de889651097ce5c1bfc3bb5f43ce9)