Update patch set 28
Patch Set 28:

Colin backwards compatibility might not be such a big problem when you can write an exploit that could get the unencrypted passwords in like 5 minutes. This really needs to be solved tbh.

Theres also a possibility to not break anyone, you could leave the old way that would use unencrypted passwords and provide a new way that actually doesnt touch the datastore before the passwords are encrypted.

Patch-set: 28