Remove <Embed-Dependency> in netconf-config pom
The pom embeds the threadpool-config-impl dependency. Not sure exactly
why this was done but I suspect as a work around for the java imports from
the blueprint XML not being included in <Import-Package>. We should not
embed the dependency. The maven-bundle-plugin can scan blueprint XML files
but there is an issue with using it so we can instead add the necessary
packages to <Import-Package>.

Change-Id: Ice33dd0e7b7f75f9d277d83ab05e6c8303867ed4
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>