Bug 8939: Extract MapEntryNode from MapNode result of XML parsing
Originally, XmlNormalizedNodeBodyReader expected a MapEntryNode
as the result of parsing an XML input with a root list element.

However, after migrating to the new XML parser, the result of parsing
such an XML input is a MapNode with one MapEntryNode inside.

Therefore, extract this MapEntryNode and use it as the result of parsing.

Change-Id: I778bf8036d7b90f75d2d5028c5a9e31ae4db995e
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>
Signed-off-by: Jakub Morvay <jmorvay@cisco.com>
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>