Update patch set 1
Patch Set 1:

Okay no worries let me take out the Service registration.  This complexity can be solved in ApiDocApplication by throwing in a WaitingServiceTracker... but in this case we can just ignore that detail for now to get this in.  I'll remove the service definition from blueprint.xml and resubmit.

Patch-set: 1